rag_slicepolicy/requirements.txt
pandas
faiss-cpu
sentence-transformers
openpyxl